Added button support for simple operations
organised the code with frames and classes. Main part now consists of 2 frames - one as an entry and the other includes useful buttons such as "x" - closes the program,  "C" - clears the entry field, "=" - same as enter key bind - clears entry field and displays an answer, "🠕" - shows another menu(frame), which consists of the numbers (0 to 10) and "🠔", which clears one character from the entry field per click. As well as that, "🠕" adds an addition, subtraction, multiplication, division and grading operators to the frame with the "x", "C"... buttons